Strukturendring før refaktorering
